+++ src/main/java/com/breixo/library/domain/port/output/book/BookDeletionPersistencePort.java
@@ -10,5 +10,5 @@ public interface BookDeletionPersistencePort {
      *
      * @param id the book identifier.
      */
-    void execute(@NotNull Long id);
+    void execute(@NotNull Integer id);
 }
