+++ src/main/java/com/breixo/library/infrastructure/adapter/input/web/controller/book/PatchBookController.java
@@ -27,7 +27,7 @@ public class PatchBookController implements PatchBookV1Api {
 
     /** {@inheritDoc} */
     @Override
-    public ResponseEntity<BookV1ResponseDto> patchBookV1(final Long id, final PatchBookV1Request patchBookV1Request) {
+    public ResponseEntity<BookV1ResponseDto> patchBookV1(final Integer id, final PatchBookV1Request patchBookV1Request) {
 
         final var updateBookCommand = this.patchBookRequestMapper.toUpdateBookCommand(id, patchBookV1Request);
