+++ src/main/java/com/breixo/library/infrastructure/adapter/output/repository/book/BookUpdatePersistenceRepository.java
@@ -55,7 +55,7 @@ public class BookUpdatePersistenceRepository implements BookUpdatePersistencePor
      * @param id the book identifier.
      * @return the book.
      */
-    private Book find(final Long id) {
+    private Book find(final Integer id) {
 
         final var bookSearchCriteriaCommand = BookSearchCriteriaCommand.builder().id(id).build();
