+++ src/test/java/com/breixo/library/infrastructure/adapter/input/web/mapper/user/PatchUserRequestMapperTest.java
@@ -34,7 +34,7 @@ class PatchUserRequestMapperTest {
     @Test
     void testToUpdateUserCommand_whenRequestIsValid_thenReturnMappedCommand() {
         // Given
-        final var id = Instancio.create(Long.class);
+        final var id = Instancio.create(Integer.class);
         final var patchUserV1RequestDto = Instancio.create(PatchUserV1Request.class);
         final var userStatus = Instancio.create(UserStatus.class);
 
@@ -86,7 +86,7 @@ class PatchUserRequestMapperTest {
     @Test
     void testToUpdateUserCommand_whenRequestIsNull_thenReturnCommandWithOnlyId() {
         // Given
-        final var id = Instancio.create(Long.class);
+        final var id = Instancio.create(Integer.class);
 
         // When
         final var updateUserCommand = this.patchUserRequestMapper.toUpdateUserCommand(id, null);
